requirements.txt
Pillow==10.4.0
black==24.8.0
fire==0.6.0
pydantic==2.9.0